+++ chanjo_report/interfaces/__init__.py
@@ -1,3 +1,0 @@
-# -*- coding: utf-8 -*-
-from .html import render_html
-from .pdf import render_pdf